perf(chat): rename virtual policy file to avoid markdown PR clash
Keep virtualize threshold / overscan knobs in chatVirtualPolicy.ts so
this branch does not fight #508's streamRenderPolicy (plain markdown).
